% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean function to calculate the Greatest Common Divisor(GCD)}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{the first numeric scalar/integer argument}

\item{b}{the second numeric scalar/integer argument}
}
\description{
Euclidean function to calculate the Greatest Common Divisor(GCD)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean}
}
