% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph data}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\usage{
wiki_graph
}
\description{
Wiki_graph is a dataframe with 3 variables.
v1 and v2 gives the edges of a graph. w gives the weight of
the particular edge in the graph.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\keyword{datasets}
