% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra function to calculate the shortest path}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{a data frame with three arguments v1,v2 and w}

\item{init_node}{source node}
}
\description{
Dijkstra function to calculate the shortest path
}
