% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group03lab03.R
\docType{package}
\name{Group03lab03}
\alias{Group03lab03}
\title{Group03lab03: A package that contains the Dijkstra
and Euclidean algorithm}
\description{
The Group03lab03 package contains two algorithms:
\enumerate{
\item dijkstra
\item euclidean
}
}
\section{Group03lab03 functions}{

\enumerate{
\item dijkstra
returns the distance from a starting node to all other nodes
\item euclidean
calculates the gcd of two numbers
}
}

